% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_datasets.R
\docType{data}
\name{human_datasets}
\alias{human_datasets}
\title{Human datasets}
\usage{
data(human_datasets)
}
\description{
A table of human datasets from various sources (e.g., genetics, imagery, omics) and accesible in open-access.
}
\keyword{datasets}
