% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainExistingDatasets-package.R
\docType{package}
\name{MainExistingDatasets-package}
\alias{MainExistingDatasets}
\alias{MainExistingDatasets-package}
\title{MainExistingDatasets: Main Existing Human Datasets}
\description{
Shiny for Open Science to visualize, share, and inventory the main existing human datasets for researchers.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Baptiste Couvy-Duchesne
}

}
\keyword{internal}
